♻️  [REFACTOR] 홈 페이지 및 설정 페이지 리팩토링
- Index.tsx 파일을 HomePage.tsx로 변경하고, 관련 컴포넌트로 분리하여 구조 개선
- SettingsPage.tsx 파일을 새로 생성하고, 계정 및 테마 설정 컴포넌트 추가
- 대기실 목록 및 실시간 대결 카드 컴포넌트 추가
- 기존 Index 및 SettingsPage 삭제
- 홈 페이지에서 사용자 정보를 효과적으로 표시하도록 개선
